+++ android/src/main/kotlin/ai/datatower/core_flutter/pigeon/impl/DtPigeonImpl.kt
@@ -1,10 +1,13 @@
 package ai.datatower.core_flutter.pigeon.impl
 import DTPigeon
 import DTLogLevel
+import DTPresetEvent
+import ai.datatower.ad.AdType
 
 import android.content.Context
 import android.util.Log
 import ai.datatower.analytics.DT
+import ai.datatower.analytics.utils.PresetEvent
 import org.json.JSONObject
 
 internal class DtPigeonImpl(private val context: Context): DTPigeon {
@@ -37,5 +40,20 @@ internal class DtPigeonImpl(private val context: Context): DTPigeon {
 
     override fun enableUpload() {
         DT.enableUpload()
+    }
+
+    private val presetEvents by lazy { PresetEvent.values().associateBy { it.ordinal } }
+    private fun DTPresetEvent.toDtType(): PresetEvent? = presetEvents[this.ordinal]
+
+    override fun enableAutoTrack(presetEvent: DTPresetEvent) {
+        presetEvent.toDtType()?.let {
+            DT.enableAutoTrack(it)
+        } ?: Log.w("DT Flutter", "[enableAutoTrack] Given PresetEvent is not recognizable: $presetEvent")
+    }
+
+    override fun disableAutoTrack(presetEvent: DTPresetEvent) {
+        presetEvent.toDtType()?.let {
+            DT.disableAutoTrack(it)
+        } ?: Log.w("DT Flutter", "[disableAutoTrack] Given PresetEvent is not recognizable: $presetEvent")
     }
 }